Extract duplicated logic for another push notification notifier to a module
