+++ app/models/push_notification_notifier.rb
@@ -1,0 +1,19 @@
+require "houston"
+
+module PushNotificationNotifier
+  private
+
+  def client
+    @client ||= Houston::Client.respond_to?(Rails.env) ?
+                  Houston::Client.send(Rails.env) :
+                  Houston::Client.new
+  end
+
+  def notifications_for(player)
+    Device.for(player).map { |device| notification(player, device.token) }
+  end
+
+  def notification(player, token)
+    raise NotImplementedError, "Class must implement the notification"
+  end
+end
